// apps/web/src/lib/shopify/types.ts
export type ShopifyImage = {
  url: string;
  altText: string | null;
  width: number;
  height: number;
};